requirements.txt
amqp==5.0.6
appnope==0.1.2
arabic-reshaper==2.1.3
asgiref==3.4.1
astroid==2.8.6
autopep8==1.6.0
backcall==0.2.0
billiard==3.6.4.0
black==21.11b1
boto3==1.20.13
botocore==1.23.13
cachetools==4.2.4
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.0
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.46.0
django-celery-beat==2.2.1
django-countries==7.2.1
django-debug-toolbar==3.2.2
django-decouple==2.1
django-filter==21.1
django-mathfilters==1.0.0
django-paypal==1.1.1
django-ses==2.3.1
django-simple-history==3.0.0
django-storages==1.12.3
django-timezone-field==4.2.1
django-user-agents==0.4.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-metadata==4.8.2
ipinfo==3.0.0
ipinfo-django==1.1.0
ipython==7.29.0
ipython-genutils==0.2.0
isort==5.10.1
jdcal==1.4.1
jedi==0.18.1
jmespath==0.10.0
kombu==5.2.2
lazy-object-proxy==1.6.0
Markdown==3.3.6
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.4
oauthlib==3.1.1
openpyxl==3.0.9
pandas==1.3.4
parso==0.8.2
pathspec==0.9.0
paystackapi==2.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prompt-toolkit==3.0.22
psycopg2==2.9.2
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
pylint==2.11.1
pyotp==2.6.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.5
python-memcached==1.59
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
reportlab==3.6.2
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
sentry-sdk==1.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
traitlets==5.1.1
twilio==7.3.1
typing_extensions==4.0.0
ua-parser==0.10.0
urllib3==1.26.7
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.13.3
zipp==3.6.0
